+++ changelog.md
@@ -6,6 +6,7 @@ This changelog helps developers and users to keep track of new features, fixes a
 
 * fix: correct escape for umlauts in configuration description
 * chore: implement against latest Jira version
+* refactor: rename Scrum poker session service implementation
 
 ## [3.3.0] - 2018-09-01